Add flag for disabling sessions
This is needed when you're bootstrapping new accounts. In that case, trying to
run IAM operations with a session will result in an `InvalidClientTokenId` error
due to MFA requirements not being met. For those cases, you need to run
aws-vault without a session so you can setup the initial IAM users, roles, etc.